pyproject.toml
[project]
name = "autovalor"
version = "0.1.0"
description = "Vehicle Valuation ML API"
requires-python = ">=3.9"
authors = [
    {name = "Your Name", email = "your.email@example.com"}
]
dependencies = [
    "scikit-learn>=1.4.0",
    "pandas>=2.1.0",
    "numpy>=1.26.0",
    "python-dotenv>=1.0.0",
    "supabase==2.10.0",
    "fastapi>=0.128.8",
    "uvicorn>=0.39.0",
    "httpx>=0.27.2",
    "playwright>=1.58.0",
    "xgboost>=2.0.0",
    "joblib>=1.3.0",
    "anthropic>=0.40.0",
    "python-multipart>=0.0.9",
    "playwright-stealth>=2.0.3",
    "requests>=2.32.5",
    "beautifulsoup4>=4.14.3",
]

[dependency-groups]
dev = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
addopts = "-v"
